requirements.txt
pyautogui
pillow
numpy
opencv-python
matplotlib